refactor: extract Untitled-agent rename into useRenameUntitledAgent hook
Replace the inline fetch + queryClient.invalidateQueries in AgentHome with
a dedicated useMutation. The mutation's mutationFn composes deriveAgentName
+ the PUT, and onSuccess handles invalidation. Callers fire mutate() and
don't care about the component lifecycle — the queryClient is app-scoped.
